Add or update plugins

+++ plugins/buddyboss-platform/bp-media/bp-media-filters.php
@@ -222,8 +222,9 @@ function bp_media_activity_comment_entry( $comment_id ) {
 		?>
 		<div class="bb-activity-media-wrap
 			<?php
-			echo esc_attr( 'bb-media-length-' . $media_template->media_count );
-			echo $media_template->media_count > 5 ? esc_attr( ' bb-media-length-more' ) : '';
+			$max_length = bb_media_get_activity_comment_max_thumb_length();
+			echo esc_attr( 'bb-media-length-' . ( $media_template->media_count > $max_length ? $max_length : $media_template->media_count ) );
+			echo $media_template->media_count > $max_length ? esc_attr( ' bb-media-length-more' ) : '';
 			?>
 			">
 				<?php
@@ -329,6 +330,7 @@ function bp_media_update_activity_media_meta( $content, $user_id, $activity_id )
 	remove_action( 'bp_groups_posted_update', 'bp_media_groups_activity_update_media_meta', 10, 4 );
 	remove_action( 'bp_activity_comment_posted', 'bp_media_activity_comments_update_media_meta', 10, 3 );
 	remove_action( 'bp_activity_comment_posted_notification_skipped', 'bp_media_activity_comments_update_media_meta', 10, 3 );
+	remove_action( 'bp_activity_posted_update', 'bb_activity_at_name_send_emails', 12, 3 );
 
 	$media_ids = bp_media_add_handler( $medias, $_POST['privacy'] );
 
@@ -336,6 +338,7 @@ function bp_media_update_activity_media_meta( $content, $user_id, $activity_id )
 	add_action( 'bp_groups_posted_update', 'bp_media_groups_activity_update_media_meta', 10, 4 );
 	add_action( 'bp_activity_comment_posted', 'bp_media_activity_comments_update_media_meta', 10, 3 );
 	add_action( 'bp_activity_comment_posted_notification_skipped', 'bp_media_activity_comments_update_media_meta', 10, 3 );
+	add_action( 'bp_activity_posted_update', 'bb_activity_at_name_send_emails', 12, 3 );
 
 	// save media meta for activity.
 	if ( ! empty( $activity_id ) ) {
@@ -560,7 +563,7 @@ function bp_media_forums_new_post_media_save( $post_id ) {
 		$media_ids = array();
 		foreach ( $medias as $media ) {
 
-			$title             = ! empty( $media['name'] ) ? $media['name'] : '';
+			$title             = ! empty( $media['name'] ) ? sanitize_text_field( wp_unslash( $media['name'] ) ) : '';
 			$attachment_id     = ! empty( $media['id'] ) ? $media['id'] : 0;
 			$attached_media_id = ! empty( $media['media_id'] ) ? $media['media_id'] : 0;
 			$album_id          = ! empty( $media['album_id'] ) ? $media['album_id'] : 0;
@@ -775,7 +778,7 @@ function bp_media_forums_embed_gif( $content, $id ) {
 				<video preload="auto" playsinline poster="<?php echo $preview_url; ?>" loop muted>
 					<source src="<?php echo $video_url; ?>" type="video/mp4">
 				</video>
-				<a href="#" class="gif-play-button">
+				<a href="#" class="gif-play-button" aria-label="<?php esc_attr_e( 'Play GIF', 'buddyboss' ); ?>">
 					<span class="bb-icon-bl bb-icon-play"></span>
 				</a>
 				<span class="gif-icon"></span>
@@ -828,7 +831,11 @@ function bp_media_forums_save_gif_data( $post_id ) {
 			bp_activity_update_meta( $main_activity_id, '_gif_data', $gdata );
 			bp_activity_update_meta( $main_activity_id, '_gif_raw_data', $gif_data );
 		}
-	} else {
+	} elseif (
+		isset( $_POST['action'] ) &&
+		in_array( $_POST['action'], array( 'bbp-edit-reply', 'bbp-edit-topic', 'bbp-edit-forum' ), true ) &&
+		empty( $_POST['bbp_media_gif'] )
+	) {
 		delete_post_meta( $post_id, '_gif_data' );
 		delete_post_meta( $post_id, '_gif_raw_data' );
 
@@ -1137,7 +1144,7 @@ function bp_media_activity_embed_gif_content( $activity_id ) {
 				<video preload="auto" playsinline poster="<?php echo $preview_url; ?>" loop muted>
 					<source src="<?php echo $video_url; ?>" type="video/mp4">
 				</video>
-				<a href="#" class="gif-play-button">
+				<a href="#" class="gif-play-button" aria-label="<?php esc_attr_e( 'Play GIF', 'buddyboss' ); ?>">
 					<span class="bb-icon-bl bb-icon-play"></span>
 				</a>
 				<span class="gif-icon"></span>
